Closes #217,#218.  Geslachten afdelingen gaan mee naar nieuw werkjaar; cache wordt geinvalideerd na jaarovergang.
git-svn-id: https://develop.chiro.be/subversion/cg2/trunk@1402 99b415e2-69d1-4072-8d93-0ed3a15484ab

+++ Solution/Chiro.Gap.WebApp/HtmlHelpers/DropDownListHelper.cs
@@ -68,7 +68,7 @@ namespace Chiro.Gap.WebApp.HtmlHelpers
 			foreach (var geslacht in Enum.GetValues(typeof(GeslachtsType)))
 			{
 				var option1 = new TagBuilder("option");
-				option1.MergeAttribute("value", geslacht.ToString());
+				option1.MergeAttribute("value", ((int)geslacht).ToString());
 				if ((GeslachtsType)geslacht == g)
 				{
 					option1.MergeAttribute("selected", @"yes");
